Refactor Watchable Registry
