+++ scrutiny/gui/core/watchable_registry/errors.py
@@ -1,0 +1,19 @@
+
+
+__all__ = [
+    'WatchableRegistryError',
+    'WatchableRegistryNodeNotFoundError',
+    'WatcherNotFoundError'
+]
+
+
+class WatchableRegistryError(Exception):
+    pass
+
+
+class WatchableRegistryNodeNotFoundError(WatchableRegistryError):
+    pass
+
+
+class WatcherNotFoundError(Exception):
+    pass
