+++ scrutiny/gui/dialogs/value_import_dialog.py
@@ -20,7 +20,8 @@ from PySide6.QtWidgets import (QDialog, QWidget, QVBoxLayout, QFormLayout, QHBox
 from PySide6.QtCore import Qt
 
 from scrutiny.gui.core.serializable_value_set import SerializableValueSet
-from scrutiny.gui.core.watchable_registry import WatchableRegistry, ParsedFullyQualifiedName, WatchableRegistryError
+from scrutiny.gui.core.watchable_registry.fqn import ParsedFullyQualifiedName, FQN
+from scrutiny.gui.core.watchable_registry.errors import WatchableRegistryError
 from scrutiny.gui.core.server_manager.server_manager import ServerManager
 from scrutiny.gui.widgets.watchable_tree import get_watchable_icon
 from scrutiny.gui.widgets.scrutiny_qmenu import ScrutinyQMenu
@@ -101,7 +102,7 @@ class PathItem(QTableWidgetItem):
     _parsed_fqn: ParsedFullyQualifiedName
 
     def __init__(self, fqn: str) -> None:
-        self._parsed_fqn = WatchableRegistry.FQN.parse(fqn)
+        self._parsed_fqn = FQN.parse(fqn)
         super().__init__(get_watchable_icon(self._parsed_fqn.watchable_type), self._parsed_fqn.path)
         self.setTextAlignment(Qt.AlignmentFlag.AlignLeft | Qt.AlignmentFlag.AlignVCenter)
         self.setToolTip(self._parsed_fqn.path)
