+++ test/gui/manual/manual_test_varlist_watch_components.py
@@ -18,7 +18,7 @@ app = make_manual_test_app()
 from PySide6.QtWidgets import QMainWindow, QLabel, QWidget, QVBoxLayout, QHBoxLayout, QPushButton, QFormLayout
 from scrutiny.gui.components.globals.varlist.varlist_component import VarListComponent
 from scrutiny.gui.components.locals.watch.watch_component import WatchComponent
-from scrutiny.gui.core.watchable_registry import WatchableRegistry
+from scrutiny.gui.core.watchable_registry.watchable_registry import WatchableRegistry
 from scrutiny.gui.component_app_interface import AbstractComponentAppInterface
 from test.gui.fake_server_manager import FakeServerManager, ServerConfig
 from scrutiny.tools.typing import *
